--- a/src/hotspot/share/opto/library_call.cpp
+++ b/src/hotspot/share/opto/library_call.cpp
@@ -5473,58 +5473,38 @@ bool LibraryCallKit::inline_array_partition() {
   Node* toIndex         = argument(5);
   Node* indexPivot1     = argument(6);
   Node* indexPivot2     = argument(7);
+  Node* pivot_indices   = argument(8);
 
-  Node* pivotIndices = nullptr;
-
-  // Set the original stack and the reexecute bit for the interpreter to reexecute
-  // the bytecode that invokes DualPivotQuicksort.partition() if deoptimization happens.
-  { PreserveReexecuteState preexecs(this);
-    jvms()->set_should_reexecute(true);
-
-    const TypeInstPtr* elem_klass = gvn().type(elementType)->isa_instptr();
-    ciType* elem_type = elem_klass->const_oop()->as_instance()->java_mirror_type();
-    BasicType bt = elem_type->basic_type();
-    // Disable the intrinsic if the CPU does not support SIMD sort
-    if (!Matcher::supports_simd_sort(bt)) {
-      return false;
-    }
-    address stubAddr = nullptr;
-    stubAddr = StubRoutines::select_array_partition_function();
-    // stub not loaded
-    if (stubAddr == nullptr) {
-      return false;
-    }
-    // get the address of the array
-    const TypeAryPtr* obj_t = _gvn.type(obj)->isa_aryptr();
-    if (obj_t == nullptr || obj_t->elem() == Type::BOTTOM ) {
-      return false; // failed input validation
-    }
-    Node* obj_adr = make_unsafe_address(obj, offset);
-
-    // create the pivotIndices array of type int and size = 2
-    Node* size = intcon(2);
-    Node* klass_node = makecon(TypeKlassPtr::make(ciTypeArrayKlass::make(T_INT)));
-    pivotIndices = new_array(klass_node, size, 0);  // no arguments to push
-    AllocateArrayNode* alloc = tightly_coupled_allocation(pivotIndices);
-    guarantee(alloc != nullptr, "created above");
-    Node* pivotIndices_adr = basic_plus_adr(pivotIndices, arrayOopDesc::base_offset_in_bytes(T_INT));
-
-    // pass the basic type enum to the stub
-    Node* elemType = intcon(bt);
-
-    // Call the stub
-    const char *stubName = "array_partition_stub";
-    make_runtime_call(RC_LEAF|RC_NO_FP, OptoRuntime::array_partition_Type(),
-                      stubAddr, stubName, TypePtr::BOTTOM,
-                      obj_adr, elemType, fromIndex, toIndex, pivotIndices_adr,
-                      indexPivot1, indexPivot2);
-
-  } // original reexecute is set back here
-
-  if (!stopped()) {
-    set_result(pivotIndices);
+  const TypeInstPtr* elem_klass = gvn().type(elementType)->isa_instptr();
+  ciType* elem_type = elem_klass->const_oop()->as_instance()->java_mirror_type();
+  BasicType bt = elem_type->basic_type();
+  // Disable the intrinsic if the CPU does not support SIMD sort
+  if (!Matcher::supports_simd_sort(bt)) {
+    return false;
   }
+  address stubAddr = StubRoutines::select_array_partition_function();
+  // stub not loaded
+  if (stubAddr == nullptr) {
+    return false;
+  }
+  // get address of the array
+  obj = must_be_not_null(obj, true);
+  const TypeAryPtr* obj_t = _gvn.type(obj)->isa_aryptr();
+  if (obj_t == nullptr || obj_t->elem() == Type::BOTTOM ) {
+    return false; // failed input validation
+  }
+  Node* obj_adr = make_unsafe_address(obj, offset);
+  Node* pivotIndices_adr = array_element_address(pivot_indices, intcon(0), T_INT);
 
+  // pass the basic type enum to the stub
+  Node* elemType = intcon(bt);
+
+  // Call the stub
+  const char *stubName = "array_partition_stub";
+  make_runtime_call(RC_LEAF|RC_NO_FP, OptoRuntime::array_partition_Type(),
+                    stubAddr, stubName, TypePtr::BOTTOM,
+                    obj_adr, elemType, fromIndex, toIndex, pivotIndices_adr,
+                    indexPivot1, indexPivot2);
   return true;
 }
 
@@ -5545,8 +5525,8 @@ bool LibraryCallKit::inline_array_sort() {
   if (!Matcher::supports_simd_sort(bt)) {
     return false;
   }
-  address stubAddr = nullptr;
-  stubAddr = StubRoutines::select_arraysort_function();
+  address stubAddr = StubRoutines::select_arraysort_function();
+
   //stub not loaded
   if (stubAddr == nullptr) {
     return false;