@mur47x111 2024-03-20
/label remove hotspot