@vamsi-parasa 2024-03-20
> @vamsi-parasa @vnkozlov I would like to get some comments on this draft PR before proceeding further. It is for simplifying the implementations of intrinsics in JIT.

Hello Yudi (@mur47x111), could you please share the performance data with and without the patch? The current Java side API was suggested so that it would be possible to enable DPQS for MemorySegments as well. Could you please explain the benefits of the Java side changes?