@mur47x111 2024-03-25
Hi Srinivas, sorry for the late reply. 

> could you please share the performance data with and without the patch?

The performance data is as follows
| Arrays.sort benchmark | Array Size | Baseline (us/op) | With patch (us/op) | Speedup |
|-----------------------|------------|------------------|---------------------|---------|
| ArraysSort.doubleSort | 10         | 0.044            | 0.044               | 1.00     |
| ArraysSort.doubleSort | 25         | 0.149            | 0.135               | 1.10     |
| ArraysSort.doubleSort | 50         | 0.343            | 0.368               | 0.93     |
| ArraysSort.doubleSort | 75         | 0.507            | 0.511               | 0.99     |
| ArraysSort.doubleSort | 100        | 0.884            | 0.848               | 1.04     |
| ArraysSort.doubleSort | 1000       | 19.647            | 16.453               | 1.19     |
| ArraysSort.doubleSort | 10000      | 437.070          | 443.245              | 0.99     |
| ArraysSort.doubleSort | 100000     | 4815.197         | 4852.364            | 0.99     |
| ArraysSort.doubleSort | 1000000    | 55425.103        | 53917.347           | 1.02     |
| ArraysSort.floatSort  | 10         | 0.053            | 0.052               | 1.02     |
| ArraysSort.floatSort  | 25         | 0.120            | 0.118               | 1.02     |
| ArraysSort.floatSort  | 50         | 0.331            | 0.346                | 0.96     |
| ArraysSort.floatSort  | 75         | 0.655            | 0.637               | 1.03     |
| ArraysSort.floatSort  | 100        | 0.734            | 0.750               | 0.98     |
| ArraysSort.floatSort  | 1000       | 17.997            | 18.969               | 0.95     |
| ArraysSort.floatSort  | 10000      | 426.431          | 440.108              | 0.97     |
| ArraysSort.floatSort  | 100000     | 4636.867          | 4621.305             | 1.00     |
| ArraysSort.floatSort  | 1000000    | 53977.675         | 55638.695            | 0.97     |
| ArraysSort.intSort    | 10         | 0.034            | 0.035               | 0.97     |
| ArraysSort.intSort    | 25         | 0.101            | 0.101               | 1.00     |
| ArraysSort.intSort    | 50         | 0.358            | 0.365               | 0.98     |
| ArraysSort.intSort    | 75         | 0.712            | 0.713               | 1.00     |
| ArraysSort.intSort    | 100        | 0.759             | 0.742               | 1.02     |
| ArraysSort.intSort    | 1000       | 19.137           | 19.592               | 0.98     |
| ArraysSort.intSort    | 10000      | 440.766          | 456.244              | 0.96     |
| ArraysSort.intSort    | 100000     | 4543.698         | 4460.939             | 1.02     |
| ArraysSort.intSort    | 1000000    | 52257.312        | 53034.259            | 0.99     |
| ArraysSort.longSort   | 10         | 0.041            | 0.041               | 1.00     |
| ArraysSort.longSort   | 25         | 0.113            | 0.117                | 0.99     |
| ArraysSort.longSort   | 50         | 0.286            | 0.293               | 0.98     |
| ArraysSort.longSort   | 75         | 0.529            | 0.522               | 1.01     |
| ArraysSort.longSort   | 100        | 0.991             | 0.950                | 1.04     |
| ArraysSort.longSort   | 1000       | 16.053           | 15.408               | 1.04     |
| ArraysSort.longSort   | 10000      | 412.635          | 437.414              | 0.94     |
| ArraysSort.longSort   | 100000     | 4565.011         | 4465.091             | 1.02     |
| ArraysSort.longSort   | 1000000    | 52370.894        | 52488.279            | 1.00     |

> Could you please explain the benefits of the Java side changes?

While implementing the intrinsics for Graal, I found it abnormal to ignore many parameters passing to these intrinsic candidates. This may be error-prone if a different unexpected parameter is passed in a future extension, because the JIT compiler will still intrinsify as per the original algorithm. 

I don't have strong opinion against keeping the `offset`, as I did not realize it was for off-heap case. It was duplicating the format of `Arrays.vectorizedHashCode`. I mainly want to push for the allocation-free change -- this will simplify the implementation in both C2 and Graal.