@vamsi-parasa 2024-03-26
Hello Yudi (@mur47x111),

Thank you for the performance data. For your internal testing, it is recommended to use the comprehensive ArraysSort.Java benchmark developed by Vladimir Yaroslavskiy. The benchmark is available at: https://github.com/openjdk/jdk/blob/4744a2603c14853bc60dd6909af0f727ea32a3b3/test/micro/org/openjdk/bench/java/util/ArraysSort.java 

We have experimented with similar Java side changes initially. Based on advice from Paul Sandoz, we made the Java side changes to make sure that DPQS can be used for MemorySegments in future. Will tag Paul into this discussion and you can seek his guidance on this matter.

Thanks,
Vamsi