@vamsi-parasa 2024-03-27
Hello Paul (@PaulSandoz), 

Could you please take a look at this PR?
Yudi (@mur47x111) is working on making changes to the Java side API for the DualPivotQuickSort.java in order to be able to support intrinsics for Graal.

These changes will undo the existing implementation which was based on your suggestion to support MemorySegments in future as well.

Thanks,
Vamsi