@mur47x111 2024-03-27
To clarify, there is no showstopper in the Java side API for Graal intrinsic implementation. We could of course do what C2 intrinsic implementation does.

As I said in the summary, there are three parts in the Java side API change:
1. makes DualPivotQuicksort.partition allocation free. Instead of returning new int[2], it now returns a long value combining both integers. This will simplify the Graal implementation (as well as C2's).
2. moves fallback implementations into intrinsic candidates instead of passing as argument. Graal can of course ignore the backup implementation argument as C2 does. I think it is less error-prone and less confusing if we move fallback implementation into the intrinsic candidate.
3. removes parameter `offset` and relies on the JIT compiler to resolve offset of the first array element. This is the key part to future MemorySegments DPQS extension. I personally don't have strong opinion to reverting this part of the change. One minor thing to consider here is whether we should disallow out-of-bound sorting on the API level.

It would be great if we could get some opinions from C2's perspective. /cc @dean-long @iwanowww