@dean-long 2024-03-28
What about something like this:

```
int[] pivotIndices = new int[2];
partition(long.class, a, Unsafe.ARRAY_LONG_BASE_OFFSET, low, high, e1, e5, DualPivotQuicksort::partitionDualPivot, pivotIndices);
lower = pivotIndices[0];
upper = pivotIndices[1];
```
where partition() is declared as:
```
int[] partition(A a, int low, int high, int pivotIndex1, int pivotIndex2, int[] pivotIndicesResult);

```
and hope escape analysis can scalarize the allocation.